Move libCEED files in a libceed subdirectory.
