+++ fem/libceed/ceed.cpp
@@ -9,8 +9,8 @@
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
 
-#include "libceed.hpp"
-#include "../general/device.hpp"
+#include "ceed.hpp"
+#include "../../general/device.hpp"
 
 namespace mfem
 {
